requirements.txt
python-telegram-bot==21.*
SQLAlchemy==2.*
python-dotenv